backend/requirements.txt
# Python 3.11+ required (tested on 3.13 with unpinned versions)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
python-whois>=0.9.4
tldextract>=5.1.2
pydantic>=2.7.1
slowapi>=0.1.9
upstash-redis>=1.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
lxml>=4.9
